requirements.txt
openai>=1.0
gradio>=5.0
sentence-transformers>=3.0
pyyaml>=6.0
python-dotenv>=1.0
